requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.16.0
transformers>=4.30.0
accelerate>=0.21.0
matplotlib>=3.7.0
numpy>=1.24.0
pillow>=9.4.0
tqdm>=4.65.0 